Traffic Sign Recognition/requirements.txt
tensorflow-gpu>=2.6.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=0.24.0
opencv-python>=4.5.0
kagglehub>=0.1.0